Restored Visual Studio build support
Wrapped code removed during Linux porting in a _WIN32 conditional block.
This could be neater but keeps it closer to the original code base.

Updated to VS2017 platform toolset, which is still statically linked.
The x86 version remains compatible with XP, and no longer uses enhanced
instructions so it's compatible with older systems.

Also fixed Open Folder support in Visual Studio, which uses CMake.
We weren't setting the required output folders correctly. The import
library (.lib) goes in lib, the DLL goes in bin, and headers are under
include/caps as normal.